requirements.txt
torch>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0
matplotlib>=3.7.0
tensorboard>=2.13.0
python-chess>=1.9.0
pygame>=2.5.0 